+++ OCP/Circle.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OCP
+{
+    public class Circle : IShape
+    {
+        public double Radius { get; set; }
+        public double CalculateArea()
+        {
+            return Math.PI * Radius * Radius;
+        }
+    }
+}
