+++ OCP/Triangle.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OCP
+{
+    public class Triangle : IShape
+    {
+        public double Base { get; set; }
+        public double Height { get; set; }
+
+        public double CalculateArea()
+        {
+            return 0.5 * Base * Height;
+        }
+    }
+}
