+++ WWSRP/Participant.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WWSRP
+{
+    public class Participant
+    {
+        //public string? Name { get; set; }
+        public string? Username { get; set; }
+        public string? Password { get; set; }
+    }
+}
